Delete ciphers from db only if the list is not empty
Fixed 500 - LP-Database-Error

+++ server/src/main/kotlin/dev/medzik/librepass/server/controllers/api/Cipher.kt
@@ -126,7 +126,9 @@ class CipherController @Autowired constructor(
         }
 
         // delete ciphers from database
-        cipherRepository.deleteAllByIdInAndOwner(request.deleted, user.id)
+        if (request.deleted.isNotEmpty()) {
+            cipherRepository.deleteAllByIdInAndOwner(request.deleted, user.id)
+        }
 
         // save ciphers into database
         if (request.updated.isNotEmpty()) {
